perftool/autoparallel/parallel_strategy_search/requirements.txt
pyyaml
numpy<2.0.0
ortools
matplotlib
chardet
highspy
